create test for check unique columns
