+++ orientdb/src/test/resources/hibernate.properties
@@ -10,8 +10,8 @@ hibernate.ogm.datastore.create_database=true
 hibernate.ogm.datastore.database=ogm_test_database
 hibernate.ogm.datastore.username=admin
 hibernate.ogm.datastore.password=admin
-#hibernate.ogm.orientdb.storage=memory
-hibernate.ogm.orientdb.storage=plocal
+hibernate.ogm.orientdb.storage=memory
+#hibernate.ogm.orientdb.storage=plocal
 hibernate.ogm.orientdb.format.datetime=yyyy-MM-dd HH:mm:ss.SSS z
 hibernate.ogm.orientdb.format.date=yyyy-MM-dd
 hibernate.ogm.orientdb.dbtype=document
